# main-services/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: cloud API Service
  version: "v0.1.0"
security:
  - BearerAuth: []
servers:
  - url: http://127.0.0.1:8080/
paths:
...
  /v1/games/sell/genre:
    get:
      summary: get total sell of each game by genre
      description:  get total sell of each game by genre
      tags:
        - Games
      operationId: GetTotalSellByGenre
      parameters:
        - name: start
          in: query
          schema:
            type: integer
        - name: end
          in: query
          schema:
            type: integer
        - name: output
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Returns games
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - properties:
                      items:
                        type: array
                        items:
                          type: integer
  /v1/games/sell/publisher:
    get:
      summary: get total sell of each game by publisher
      description: get total sell of each game by publisher
      tags:
        - Games
      security:
        - { }
      operationId: GetTotalSellByPublisher
      parameters:
        - name: start
          in: query
          schema:
            type: integer
        - name: end
          in: query
          schema:
            type: integer
        - name: publisher1
          in: query
          schema:
            type: string
        - name: publisher2
          in: query
          schema:
            type: string
        - name: output
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Returns games
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - properties:
                      items:
                        type: array
                        items:
                          type: integer
  /v1/games/sell/year:
    get:
      summary: get total sell of all game by year
      description: get total sell of all game by year
      tags:
        - Games
      operationId: GetTotalSellByYears
      parameters:
        - name: start
          in: query
          schema:
            type: integer
        - name: end
          in: query
          schema:
            type: integer
        - name: output
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Returns games
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - properties:
                      items:
                        type: array
                        items:
                          type: integer
  /v1/games/sell/name:
    get:
      summary: get total sell of each game by name
      description: get total sell of each game by name
      tags:
        - Games
      operationId: GetTotalSellByName
      parameters:
        - name: game1
          in: query
          schema:
            type: string
        - name: game2
          in: query
          schema:
            type: string
        - name: output
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Returns games
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - properties:
                      items:
                        type: array
                        items:
                          type: integer
...
components:
  schemas:
...
    PageMeta:
      type: object
      properties:
        totalCount:
          type: integer
          format: int64
          minimum: 0
      required:
        - totalCount
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshAuth:
      type: apiKey
      in: cookie
      name: REFRESH_TOKEN
...